Update to provide an implementation of a attribute that where the data is stored on disk using a hdf file. Currently, this is pig slow for real world usage but will implemented a caching idea and hopefully within the 'real' world that should fix the performance issue.

+++ src/rastergis/RSGISFindClumpNeighbours.cpp
@@ -29,9 +29,9 @@ namespace rsgis{namespace rastergis{
         
     }
     
-    vector<list<unsigned long>* >* RSGISFindClumpNeighbours::findNeighbours(GDALDataset *clumpImage) throw(RSGISImageCalcException)
+    vector<list<size_t>* >* RSGISFindClumpNeighbours::findNeighbours(GDALDataset *clumpImage) throw(RSGISImageCalcException)
     {
-        vector<list<unsigned long>* > *neighbours = new vector<list<unsigned long>* >();
+        vector<list<size_t>* > *neighbours = new vector<list<size_t>* >();
         
         try
         {
